Passengers logic for requesting a ride
This method raises exceptions if the Passenger is not in a valid state
to request a ride (they're already in an ongoing trip).

Otherwise, a new Trip is created with this Passenger and a random
available Driver. Nil is returned if no Drivers could be found to
fulfill the request.